requirements.txt
amqp==5.0.2
aniso8601==7.0.0
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.0.5
django-cors-headers==3.5.0
django-filter==2.4.0
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.0
django-graphql-social-auth==0.1.4
django-redis==4.12.1
django-versatileimagefield==2.0
djongo==1.3.3
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
kombu==5.0.2
oauthlib==3.1.0
Pillow==8.0.1
promise==2.3
prompt-toolkit==3.0.8
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.1
python-decouple==3.3
python-magic==0.4.18
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.2.4
Unidecode==1.1.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
